User story report
Here's a report that I did with the user stories that we have in use. However, I did not put all of the issues since some of them tend to repeat themselves, and not all of them have someone assigned to them, so this is subject to change. Let me know what you think and if there's anything missing or anything to remove I'll fix it.